Changed the look of the img via container
I had a problem of transforming the image i used to a circle. Thnaks to the freecodecamp.org community i figured out that it would be better to fit the img into a container. Furthermore do squares create circles but rectangles create ovals so i set the height and width to an equal amount as you can see.